pyproject.toml
[project]
name = "lolrmm-artifacts"
version = "0.1.0"
description = "Export RMM tool artifacts from LOLRMM for SIEM/EDR detection engineering."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "jstym" }]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.7",
    "pyyaml>=6.0.1",
    "httpx>=0.27",
    "rich>=13.7",
]

[project.scripts]
lolrmm = "lolrmm_artifacts.cli:app"

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-httpx>=0.30", "ruff>=0.4"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lolrmm_artifacts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
